Fix logging header patching for svt-hevc-sys
The patched header includes EbApi.h and EbApiVersion.h, which we would normally
not patch. However, CMake always does a preflight compilation check, which with
our -include directives fails unless those files are on the -Ipath. And at the
time that preflight check is done, CMake hasn't had a chance to sed in the API
version into EbApiVersion.h, which is in the source tree.

We're already patching one file, so we may as well generate a fixed
EbApiVersion.h as well.